refactor: split query regression runner into modules
